+++ threedigrid/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
+import pkg_resources
 # Always try to import netCDF4.Dataset
-# before importing h5py to prevent netCDF4 
+# before importing h5py to prevent netCDF4
 # HDF errors.
 try:
     from netCDF4 import Dataset
@@ -10,4 +11,9 @@ except ImportError:
 
 from admin.gridadmin import GridH5Admin
 
-__version__ = '0.1.4.dev0'
+try:
+    __version__ = pkg_resources.get_distribution("threedigrid").version
+except pkg_resources.DistributionNotFound:
+    # for development environments
+    __version__ = pkg_resources.find_distributions(
+        '.', only=True).next().version